For Home Pg, I was trying to place the hat img on the top, but it seems not working somehow, and same as the
MESSAGEBOARD, I tried to adjust the margin to the right, it is not working either? I tired to make a link to the page outside of the "container", but it doesnt stay on the right side.

Somehow, the dashlines are missing on the top of the page?

I dropped 2 IMG: 1) NewArrival Logo 2) Messsageboard